Reorganised project structure to cope with python's first come first served approach to module name spaces.
Realised that python doesn't code with two packages having the same path prefix, but different source roots within a project, because the module is defined by the presence of the __init__.py file.  So, a file src/main/python/modulea/ could get masked by src/test/python/modulea/ or vice-versa, depending which is found on the path first because loading the module is defined by executing the __init__.py file in the directory, even though you may actually just want to use a package structure to organise your code and put nothing in the __init__.py file.